requirements/base.txt
django==2.0.6
djangorestframework==3.8.2
django-extensions==2.0.7
djangorestframework-jwt==1.11.0
django-filter==2.0.0
